moved pulse-specific options/solver classes into pulse/controllers
