+++ qiskit/providers/aer/pulse/controllers/pulse_de_options.py
