+++ qiskit/providers/aer/pulse/controllers/unitary_controller.py
@@ -20,9 +20,9 @@ Controller for solving unitary evolution of a state-vector.
 import time
 import numpy as np
 from scipy.linalg.blas import get_blas_funcs
-from ..de.pulse_de_options import OPoptions
+from .pulse_de_options import OPoptions
 from qiskit.tools.parallel import parallel_map, CPU_COUNT
-from ..de.pulse_de_solver import construct_pulse_zvode_solver
+from .pulse_de_solver import construct_pulse_zvode_solver
 
 # Imports from qutip_extra_lite
 from ..de.pulse_utils import occ_probabilities, write_shots_memory
